requirements.txt
Flask==2.3.2
pycryptodome==3.18.0
